% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_query}
\alias{data_query}
\title{Perform query using Data Query API}
\usage{
data_query(token, data_type = "elisa", filter = list(studyAccession =
  c("SDY2", "SDY4")))
}
\arguments{
\item{token}{str representing the token generated by the gen_token function}

\item{data_type}{string representing the data type of interest to query from API}

\item{filter}{list where each name is one ore more of filters to API and filter variables; those filters with multiple variables expressed as a vector or list}
}
\description{
This function allows the user to use token from gen_token function  to query the ImmPort API
}
\examples{
data_query(token, data_type = "elisa", filter = list(studyAccession = c("SDY2", "SDY4"), maxSubjectAgeGte=90))
}
\keyword{query}
