% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gen_token}
\alias{gen_token}
\title{Generate token function}
\usage{
gen_token(Username, Password)
}
\arguments{
\item{Username}{Your ImmPort username}

\item{Password}{Your ImmPort password}
}
\description{
This function allows the user to specify their ImmPort Username and Password to generate an API token.
}
\examples{
gen_token(Username = your_user_name, Password = your_password)
}
\keyword{token}
