% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{proc_response}
\alias{proc_response}
\title{Process contents of data query respone into data frame}
\usage{
proc_response(data_query_response)
}
\arguments{
\item{data_query_response}{the http response output from the function data_query()}
}
\description{
This function processes httr response from API into a data frame
}
\examples{
proc_response(data_query_response)
}
\keyword{data}
\keyword{frame}
