% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{show_data_types}
\alias{show_data_types}
\title{Show data types}
\usage{
show_data_types()
}
\description{
This function shows a table of data types the API can query
}
\examples{
show_data_types()
}
\keyword{api}
\keyword{data,}
\keyword{types,}
