% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{show_filters}
\alias{show_filters}
\title{Show query filters}
\usage{
show_filters()
}
\description{
This function shows a table of filters available for API
}
\examples{
show_filters()
}
\keyword{api}
\keyword{filter,}
\keyword{query,}
